Migrate to Gradle build tool
