+++ src/main/java/imsam/ProbabilityDistribution.java
@@ -1,3 +1,5 @@
+package imsam;
+
 public interface ProbabilityDistribution {
 
     /**
