+++ src/main/java/imsam/SeedPath.is
