+++ src/main/java/imsam/SparseModelGenerator.java
@@ -1,3 +1,5 @@
+package imsam;
+
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
